+++ processor/src/test/resources/expected/com/example/prop/React4j_ComponentWithArezProp.java
@@ -1,0 +1,45 @@
+package com.example.prop;
+
+import javax.annotation.Generated;
+import javax.annotation.Nonnull;
+import javax.annotation.Nullable;
+import jsinterop.base.Js;
+import jsinterop.base.JsPropertyMap;
+import react4j.core.ComponentConstructorFunction;
+import react4j.core.NativeAdapterComponent;
+import react4j.core.ReactConfig;
+
+@Generated("react4j.processor.ReactProcessor")
+class React4j_ComponentWithArezProp extends ComponentWithArezProp {
+  static final ComponentConstructorFunction TYPE = getConstructorFunction();
+
+  @Nonnull
+  private static ComponentConstructorFunction getConstructorFunction() {
+    final ComponentConstructorFunction componentConstructor = NativeReactComponent::new;
+    if ( ReactConfig.enableComponentNames() ) {
+      Js.asPropertyMap( componentConstructor ).set( "displayName", "ComponentWithArezProp" );
+    }
+    return componentConstructor;
+  }
+
+  @Override
+  protected String getValue() {
+    return props().getAny( "value" ).asString();
+  }
+
+  @Override
+  protected ComponentWithArezProp.Model getModel() {
+    return props().getAny( "model" ).cast();
+  }
+
+  private static final class NativeReactComponent extends NativeAdapterComponent<ComponentWithArezProp> {
+    NativeReactComponent(@Nullable final JsPropertyMap<Object> props) {
+      super( props );
+    }
+
+    @Override
+    protected ComponentWithArezProp createComponent() {
+      return new React4j_ComponentWithArezProp();
+    }
+  }
+}